% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_koboloadeR.R
\docType{data}
\name{data_koboloadeR}
\alias{data_koboloadeR}
\title{A raw household-level dataset with multiple type of questions}
\format{A data frame with 2407 rows and 128 variables corresponding to a relatively classic data collection ready XLSform. The form and Excel data set can be found in the "Code" folder.}
\source{
\url{http://www.reachresourcecentre.info/system/files/resource-documents/afg_dataset_emergency_response_mechanism_heat_august_2017_to_april_2018.xlsx}
}
\usage{
data_koboloadeR
}
\description{
A raw household-level dataset with multiple type of questions
}
\keyword{datasets}
