% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{format_si}
\alias{format_si}
\title{Helper function to format a vector of strings using
  SI prefix notation


Format a vector of numeric values according
to the International System of Units.
http://en.wikipedia.org/wiki/SI_prefix

Based on code by Ben Tupper
https://stat.ethz.ch/pipermail/r-help/2012-January/299804.html
Args:}
\usage{
format_si(...)
}
\arguments{
\item{\dots}{List of integer or numeric
...: Args passed to format()}
}
\value{
Formatted number.
}
\examples{
format_si()

}
\author{
Someone
}

