% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{get_me}
\alias{get_me}
\title{Helper function for GET, depending on whether authentication is required}
\usage{
get_me(user, URL)
}
\arguments{
\item{user}{string of length 1 or 2 with user details}

\item{URL}{The URL to be passed to curl}
}
\description{
Adds basic level authentication if provided.
}
\note{
This function is not intended to be called directly. It is used in other functions.
}
\author{
Ananda Mahto
}

