% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{host}
\alias{host}
\title{Specifies the host URL of the API to use}
\usage{
host(instring)
}
\arguments{
\item{instring}{Either "kobo", "kobohr", "ona", or a custom (full) URL.}
}
\value{
A single string with the URL to use.
}
\description{
A helper function to conveniently switch different APIs.
}
\note{
API URLs are made available for KoBo Toolbox ("kobo", \url{https://kc.kobotoolbox.org/api/v1/}), KoBo Humanitarian Response ("kobohr", \url{https://kc.humanitarianresponse.info/api/v1/}), and Ona ("ona", \url{https://ona.io/api/v1/}). For your own installation, or other installations using the same API but accessed at a different URL, enter the full URL.

This function is not intended to be called directly. It is used in other functions.
}
\examples{
host("kobo")
host("https://ona.io/api/v1/") ## same as host("ona")

}
\author{
Ananda Mahto
}

