% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_aggregate.R
\name{kobo_aggregate}
\alias{kobo_aggregate}
\title{Generate an aggregaation of the variable of a dataset for instance by admin level}
\usage{
kobo_aggregate(form = "form.xls", aggregVar = "admin1")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{aggregVar}{variable to use for aggregation}
}
\value{
No return, All results will be saved in the data folder
}
\description{
Generate an aggregaation of the variable of a dataset for instance by admin level
}
\examples{
\dontrun{
kobo_aggregate(form, mapref)
}

}
\author{
Edouard Legoupil
}
