% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_anonymisation_report.R
\name{kobo_anonymisation_report}
\alias{kobo_anonymisation_report}
\title{Generate anonymised dataset}
\usage{
kobo_anonymisation_report(frame, dico)
}
\arguments{
\item{dico}{Generated from kobo_dico function}

\item{kobo}{or odk dataset to use}
}
\description{
Automatically produce a disclosure risk measurement report.

 The report is generated from functions released within sdcmicro package from the worldbank.
  https://cran.r-project.org/web/packages/sdcMicro/sdcMicro.pdf
}
\examples{
kobo_anonymise()

\dontrun{
kobo_anonymisation_report(frame, dico)
}

}
\author{
Edouard Legoupil
}
