% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{kobo_apps}
\alias{kobo_apps}
\title{Shiny Apps for Viewing Online KoBo Data}
\usage{
kobo_apps(app)
}
\arguments{
\item{app}{The name of the app to be run. If empty, the function will display
the names of the available apps.}
}
\value{
Launches RStudio's viewer to view the data. The dataset is also
downloaded to your Global Environment.
}
\description{
A launcher for the Shiny apps available in the koboloadeR package.
}
\section{Available Apps}{


\itemize{
 \item \code{"data_viewer"}
}
}
\examples{

\dontrun{
kobo_apps()
kobo_apps("data_viewer")
}

}
\author{
Ananda Mahto
}

