% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_atlas_report.R
\name{kobo_atlas_report}
\alias{kobo_atlas_report}
\title{Generate an atlas out of the dataset}
\usage{
kobo_atlas_report(frame, dico, mappoly)
}
\arguments{
\item{dico}{Generated from kobo_dico function}

\item{kobo}{or odk dataset to use}

\item{map}{equaly mappoly or mappoint depending on the type of visualisation expected - polygons or points}
}
\description{
Generate report with data aggregated by location & spatial visualisation / cartography
}
\examples{
kobo_atlas_report(frame, dico, mappoly)

\dontrun{
kobo_atlas_report(frame, dico, mappoly)
}

}
\author{
Edouard Legoupil
}
