% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_multi_facet.R
\name{kobo_bar_multi_facet}
\alias{kobo_bar_multi_facet}
\title{Generate histogramm plots based on dates}
\usage{
kobo_bar_multi_facet(data, dico)
}
\arguments{
\item{data}{kobodatset to use}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate maps for all nominal & ordinal variables based on dates. ggplot2 is used.
}
\examples{
kobo_bar_multi_facet()

}
\author{
Edouard Legoupil
}

