% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunch.R
\name{kobo_bar_multi_indiv}
\alias{kobo_bar_multi_indiv}
\title{Generate bar Chart - frequency - for select_one question}
\usage{
kobo_bar_multi_indiv(question, mainDir = "")
}
\arguments{
\item{question}{Question to be treated by the function. It should be the "fullname" (as a string) of the question as generated by kobo_dico().}

\item{mainDir}{Path to the project's working directory: mainly for shiny app}
}
\value{
Returns a list containing the question fullname, ggplot object of the plot, and a frequency table, weighted if this is included in the 0-config file.
}
\description{
Generate basic data exploration analysis for kobo select_one question. Returns a list containing the question fullname, ggplot object of the plot, and a frequency table, weighted if this is included in the 0-config file. If disaggregation is set into the form, then the graph are "faceted" per disaggregation.
}
\examples{
kobo_bar_multi_indiv("S5_FOOD_ASSESS.q3_food_coping_strat")
\donttest{
kobo_bar_multi_indiv("S5_FOOD_ASSESS.q3_food_coping_strat")
}

}
\author{
Edouard Legoupil, Elliott Messeiller
}
