% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_multi_print.R
\name{kobo_bar_multi_print}
\alias{kobo_bar_multi_print}
\title{Generate bar Chart - frequency - for select_multiple questions and save output as svg for illustrator}
\usage{
kobo_bar_multi_print(data, dico)
}
\arguments{
\item{data}{kobodatset to use}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate bar chart for each of the select_multiple question in the dataset. used in report
}
\examples{
kobo_bar_multi_print()

\dontrun{
kobo_bar_multi_print(data, dico)
}


}
\author{
Edouard Legoupil
}
