% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_one_facet_print.R
\name{kobo_bar_one_facet_print}
\alias{kobo_bar_one_facet_print}
\title{Generate faceted frequency bar chart  and save output as svg for illustrator}
\usage{
kobo_bar_one_facet_print(data, dico)
}
\arguments{
\item{data}{kobodatset to use}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate faceted chart for select one variable.. ggplot2 is used.
}
\examples{
kobo_bar_one_facet_print()

\dontrun{
kobo_bar_one_facet_print(data,  dico)
}


}
\author{
Edouard Legoupil
}
