% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_one_print.R
\name{kobo_bar_one_print}
\alias{kobo_bar_one_print}
\title{Generate bar Chart - frequency - for select_one questions and save output as svg for illustrator}
\usage{
kobo_bar_one_print(data, dico)
}
\arguments{
\item{data}{.}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate bar chart for each of the select_one question in the dataset. Used in report
}
\examples{
kobo_bar_one_print()

\dontrun{
kobo_bar_one_print(data, dico)
}


}
\author{
Edouard Legoupil
}
