% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_chart_mosaic.R
\name{kobo_chart_mosaic}
\alias{kobo_chart_mosaic}
\title{Generate an association chart}
\usage{
kobo_chart_mosaic(data, .x, .y)
}
\arguments{
\item{data}{dataframe}

\item{x}{variable to use for association}

\item{y}{variable to use for association}
}
\value{
ggplot2 object
}
\description{
Generate an association chart
}
\examples{
\dontrun{
kobo_chart_mosaic(form, mapref)
}

}
\author{
Edouard Legoupil
}
