% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_check_project_configuration.R
\name{kobo_check_project_configuration}
\alias{kobo_check_project_configuration}
\title{Check kobo_check_project_configuration}
\usage{
kobo_check_project_configuration(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}
}
\value{
The return will be a list that contains a list that checks all elements of the project configuration and a message of confirmation
}
\description{
Check if the project configurations exist and all required files are in the right place.
}
\examples{
\dontrun{
kobo_check_project_configuration("myform.xls")
}

}
\author{
Maher Daoud
}
