% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_clean.R
\name{kobo_clean}
\alias{kobo_clean}
\title{Add cleaned variables to the frame based on a reference table}
\usage{
kobo_clean(frame, form = "form.xls", app = "console")
}
\arguments{
\item{frame}{dataframe to use}

\item{form}{name of the form file in xls format}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\description{
The function works on a loop based on the dictionnary.
Add a column clean and insert in the cell the name of the csv file that will be used to generate the cleaned variable
The first column of that file will be used for the matching, the second column will be added to the dataframe.
The new cleaned variable will be inserted in the dictionnary. with a suffix '.clean'
}
\examples{
\dontrun{
kobo_clean(frame, form = "form.xls")
}


}
\author{
Edouard Legoupil
}
