% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_consolidateone.R
\name{kobo_consolidateone}
\alias{kobo_consolidateone}
\title{Merge disagregated select_one variable}
\usage{
kobo_consolidateone(frame, form = "form.xls")
}
\arguments{
\item{frame}{kobo or odk exported dataset to use}

\item{form}{The full filename of the form to be accessed (has to be an xls file).}
}
\value{
data
}
\description{
Merge disagregated select_one variable
}
\examples{
\dontrun{
kobo_consolidateone(frame , form = "form.xls")
}


}
\author{
Edouard Legoupil
}
