% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_correlation_analysis.R
\name{kobo_correlation_analysis}
\alias{kobo_correlation_analysis}
\title{Correlation Analysis}
\usage{
kobo_correlation_analysis(form = "form.xls", frame, target, app = "console")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{frame}{The dataframe that contains the target variable and the independent variable(s)}

\item{target}{The name of dependent variable, the variable being tested and measured}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\value{
A list that includes all analysis and charts
}
\description{
This function apllay all correlations test to discover if there is a relation between the targe variable and other variables
}
\examples{
\dontrun{
kobo_correlation_analysis()
}

}
\author{
Maher Daoud
}
