% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_corrplot.R
\name{kobo_corrplot}
\alias{kobo_corrplot}
\title{text Could}
\usage{
kobo_corrplot(data, dico)
}
\arguments{
\item{data}{.}

\item{dico}{( generated from kobo_dico)}
}
\description{
Produce Correlation plot for all categorical vairable in the data set.
}
\examples{
kobo_corrplot()

\dontrun{
kobo_corrplot("myform.xls")
}

}
\author{
Edouard Legoupil
}
