% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_create_indicators.R
\name{kobo_create_indicators}
\alias{kobo_create_indicators}
\title{Create Indicators}
\usage{
kobo_create_indicators(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data-raw folder.}
}
\value{
No return, all results will be saved inside new CSV files
}
\description{
Function to compute indicators from indicator sheet
}
\examples{
\dontrun{
kobo_create_indicators("myform.xls")
}

}
\author{
Edouard Legoupil, Maher Daoud
}
