% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunch.R
\name{kobo_crunching}
\alias{kobo_crunching}
\title{Pick the right function for each question}
\usage{
kobo_crunching(question, mainDir = "")
}
\arguments{
\item{question}{Question to be treated by the function. It should be the "fullname" (as a string) of the question as generated by kobo_dico().}

\item{mainDir}{Path to the project's working directory: mainly for shiny app}
}
\value{
Return the result of the function.
}
\description{
This function choose and run the right type of koboloadeR function for data crunching.
}
\examples{
kobo_crunching("question_name")
\dontrun{
kobo_crunching("question_name")
}

}
\author{
Elliott Messeiller
}
