% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_data_downloader}
\alias{kobo_data_downloader}
\title{Retrieve the Data from a Specified Dataset}
\usage{
kobo_data_downloader(formid, user = NULL, api = "kobo", check = TRUE)
}
\arguments{
\item{formid}{The ID of the form to be accessed (as a character string).}

\item{user}{Optional. A single string indicating the username and password ("username:password"), or a character vector or list, length 2, with the first value being the "username", and the second being the "password".}

\item{api}{The URL at which the API can be accessed. Defaults to "kobo", which loads the KoBo Toolbox API.}

\item{check}{Logical. Should the function first check to see whether the data is available offline.}
}
\value{
A "data.table" with the full dataset. If data is already found on disk and the number of rows matches with the online datasets, the local copy would be used. The dataset would be named in the form of "data_formid".
}
\description{
Retrieves the data submitted to a specified dataset.
}
\examples{
kobo_data_downloader("15051")
kobo_data_downloader("31511", api = "kobohr")

}
\author{
Ananda Mahto
}

