% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboproj.R
\name{kobo_datasets2}
\alias{kobo_datasets2}
\title{Lists the Datasets Available including count of submission.}
\usage{
kobo_datasets2(user = NULL, api = "unhcr")
}
\arguments{
\item{user}{Optional. A single string indicating the username and password
(in the form of \code{"username:password"}), or a character vector or list,
length 2, with the first value being the "username", and the second being
the "password".}

\item{api}{The URL at which the API can be accessed. Defaults to "kobo",
which loads the KoBo Toolbox API.}
}
\value{
A data.table containing details about the datasets available,
including items like the "title", "id", and "submission".
}
\examples{
kobo_datasets2()

}
\author{
Edouard Legoupil
}

