% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_ddi.R
\name{kobo_ddi}
\alias{kobo_ddi}
\title{DDI generation}
\usage{
kobo_ddi(form = "form.xls", app = "console")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\value{
DDI version 2.5, XML file structure will saved under out/ddi
}
\description{
This function creates a DDI version 2.5, XML file structure for microdata library
}
\examples{
kobo_ddi()

\dontrun{
kobo_ddi("myform.xls")
}

}
\author{
Maher Daoud
}
