% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dico.R
\name{kobo_dico}
\alias{kobo_dico}
\title{Create Data dictionnary an the xlsform}
\usage{
kobo_dico(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}
}
\value{
A "data.frame" with the full data dictionnary. To be used in the rest of the analysis.
}
\description{
Produce a data dictionnary based on the xlsform for the project
}
\examples{
\dontrun{
kobo_dico(form = "form.xls")
}

}
\author{
Edouard Legoupil
}
