% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dummy.R
\name{kobo_dummy}
\alias{kobo_dummy}
\title{Create a dummy dataset}
\usage{
kobo_dummy(form = "form.xls")
}
\arguments{
\item{form}{file xlsform}
}
\description{
Automatically produce an dummy dataset in line with the structure of an xlsform.

Making decisions about research design and analysis strategies is often difficult before data is collected,
because it is hard to imagine the exact form data will take.
This function helps imagine what data will look like before they collect it.
samplesize is set per defautl at 500 records

 Supported Features

- Generate a data set with an output similar to the one needed in koboloader
- respects ODK structure "`relevant`" skip logic (Some advanced functionality such as "coalesce()" not covered)
  "`constraint`" and "`repeat`"
- adds InstandID column to link hierearchical data based on "`repeat_count`"
}
\examples{
\dontrun{
kobo_dummy(form)
}

}
\author{
Edouard Legoupil
}
