% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_edit_form.R
\name{kobo_edit_form}
\alias{kobo_edit_form}
\title{Edit XLS form with shiny app for configuration}
\usage{
kobo_edit_form(form = "form.xls", survey = NULL, choices = NULL,
  indicator = NULL, settings = NULL, analysisSettings = NULL)
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{survey}{Dataframe that represent the data of survey sheet in the xlsform}

\item{choices}{Dataframe that represent the data of choices sheet in the xlsform}

\item{indicator}{Dataframe that represent the data of indicator sheet in the xlsform}

\item{settings}{Dataframe that represent the data of settings sheet in the xlsform}

\item{analysisSettings}{Dataframe that represent the data of analysisSettings sheet in the xlsform}
}
\value{
No return, this function edit the original XLSform directly
}
\description{
This function used to change the data of sheets in the xlsform and apply all required styles for each sheet
}
\examples{
kobo_edit_form()

\dontrun{
kobo_edit_form("myform.xls")
}

}
\author{
Maher Daoud
}
