% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_edit_ridl_metadata.R
\name{kobo_edit_ridl_metadata}
\alias{kobo_edit_ridl_metadata}
\title{Edit RIDL / CKAN metadata}
\usage{
kobo_edit_ridl_metadata(form = "data/form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls file).
It is assumed that the form is stored in the data folder.}
}
\description{
Edit RIDL / CKAN metadata
}
\examples{
\dontrun{
kobo_edit_ridl_metadata(form = "form.xls")
}

}
\author{
Hisham Galal
}
