% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_get_begin_repeat.R
\name{kobo_get_begin_repeat}
\alias{kobo_get_begin_repeat}
\title{Get all begin repeat from xlsform}
\usage{
kobo_get_begin_repeat(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}
}
\value{
return a list that contains 1.vector of string represent the names 2.message about status
}
\description{
Get the 'name' column for all rows that have 'begin repeat' value in 'type' column
}
\examples{
\dontrun{
kobo_get_begin_repeat("myform.xls")
}

}
\author{
Maher Daoud
}
