% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_get_config.R
\name{kobo_get_config}
\alias{kobo_get_config}
\title{Get Configuration}
\usage{
kobo_get_config(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file). where settings sheet contains all configuration of the project}
}
\value{
Return a dataframe that contains configuration of the project
}
\description{
Return all configuration from Analysis Settings sheet of xlsform
}
\examples{
kobo_get_config()

}
\author{
Maher Daoud
}
