% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_get_dataframes_levels.R
\name{kobo_get_dataframes_levels}
\alias{kobo_get_dataframes_levels}
\title{Dataframes Levels}
\usage{
kobo_get_dataframes_levels(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}
}
\value{
A "data.frame" contains levels and parents for the main dataframe and all sub datasets.
}
\description{
Produce a dataframe that represents levels and parents for the main dataframe and all sub datasets.
}
\examples{
kobo_get_dataframes_levels()

\dontrun{
kobo_get_dataframes_levels("myform.xls")
}

}
\author{
Maher Daoud
}
