% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_get_theme.R
\name{kobo_get_theme}
\alias{kobo_get_theme}
\title{Get Themes}
\usage{
kobo_get_theme(form = "form.xls")
}
\value{
Return a list that contains all themes
}
\description{
Return all themes that are used for styling reports and charts
}
\examples{
kobo_get_theme()

}
\author{
Maher Daoud
}
