% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_getMainDirectory.R
\name{kobo_getMainDirectory}
\alias{kobo_getMainDirectory}
\title{get Main Directory for a KoboloadeR project}
\usage{
kobo_getMainDirectory()
}
\value{
A string for Main Directory path.
}
\description{
the function return the Main Directory for KoboloadeR packag
}
\examples{
kobo_projectinit()

}
\author{
Maher Daoud
}
