% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunch.R
\name{kobo_histo_indiv}
\alias{kobo_histo_indiv}
\title{Generate histograme for individual integer questions}
\usage{
kobo_histo_indiv(question, mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for proper shiny app path}
}
\description{
Automatically generate histogrammes for each of the integer questions in the dataset. ggplot2 is used.
}
\examples{
kobo_histo_indiv()
\dontrun{
kobo_histo_indiv()
}

}
\author{
Edouard Legoupil, Elliott Messeiller
}
