% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_indicator.R
\name{kobo_indicator}
\alias{kobo_indicator}
\title{Import & perform the indicator calculation from the XLS form}
\usage{
kobo_indicator(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for shiny app}
}
\value{
A file with all elements to get your data & form.
}
\description{
Add additional variables based on the data analysis plan to the data frame
}
\examples{
kobo_indicator()

}
\author{
Elliott Messeiller
}
