% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_left_align.R
\name{kobo_left_align}
\alias{kobo_left_align}
\title{UNHCR ggplot2 theme}
\usage{
kobo_left_align(plot_name, pieces)
}
\arguments{
\item{plot_name}{ggplot2 object}

\item{pieces}{plot labels to align}
}
\value{
Return better chart
}
\description{
Left align chart title and subtitle on a ggplot2
}
\examples{
\dontrun{
kobo_left_align()
}

}
\author{
Edouard Legoupil - with inspiration from bbc
}
