% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_load_packages.R
\name{kobo_load_packages}
\alias{kobo_load_packages}
\title{Load Packages}
\usage{
kobo_load_packages()
}
\value{
No return
}
\description{
Load all necessary packages for koboloadeR
}
\examples{
\dontrun{
kobo_load_packages()
}


}
\author{
Edouard Legoupil, Maher Daoud
}
