% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_prediction_report.R
\name{kobo_prediction_report}
\alias{kobo_prediction_report}
\title{Generate prediction}
\usage{
kobo_prediction_report(origin, location)
}
\arguments{
\item{kobo}{or odk dataset to use}
}
\description{
Automatically produce prediction reports
containing elements of descriptive analysis,
information about the methods used
(such as regression and classification) and prediction results.
The configuration of the function is done in an xlsform.
The target variables are taken from the household surveys
and three types of outcome will be considered:
binary (going to school, marry children,...),
numeric (expenditure per capita,…)
and ordinal (level of agreement, support from local community,…).
Then the function will train different models
(using the inner joint between registration and household surveys as training data)
and select the most accurate one.
The user needs to have previously loaded the proGres, household and form files from the 0-config script.
The function takes as arguments the names of the variables for origin and location of asylum, to be able
to describe the predicted outcome.
}
\author{
Damien Seite
}
