% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_prepare_form.R
\name{kobo_prepare_form}
\alias{kobo_prepare_form}
\title{Prepare XLS form}
\usage{
kobo_prepare_form(form = "form.xls")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}
}
\value{
No return, this function edit the original XLSform directly
}
\description{
Prepare XLSform by adding chapter, disaggregation, correlate, variable, anonymise, structuralequation,
clean, cluster, predict, mappoint, mappoly in case if those fields are not exist; the function will create dummy column for each one.
Also, coloring all rows that have type equal to "begin group", "end group", "begin repeat" or "end repeat".
}
\examples{
kobo_prepare_form()

\dontrun{
kobo_prepare_form("myform.xls")
}

}
\author{
Maher Daoud
}
