% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboproj.R
\name{kobo_projectinit}
\alias{kobo_projectinit}
\title{Create analysis project structure & related analysis scripts}
\usage{
kobo_projectinit()
}
\value{
A structure of directory and scripts in order to set up quickly a project.
}
\examples{
kobo_projectinit()

}
\author{
Edouard Legoupil
}

