% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_rename_xlsform_dataframes.R
\name{kobo_rename_xlsform_dataframes}
\alias{kobo_rename_xlsform_dataframes}
\title{Rename xlsform and all dataframes}
\usage{
kobo_rename_xlsform_dataframes(form = "form.xls", app = "console")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\value{
no return. only if there is error.
}
\description{
Rename xlsform under data file to form.xls and all dataframes to the
}
\examples{
\dontrun{
kobo_rename_xlsform_dataframes()
}

}
\author{
Maher Daoud
}
