% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_shiny.R
\name{kobo_shiny}
\alias{kobo_shiny}
\title{Shiny app launcher}
\usage{
kobo_shiny(app = "")
}
\arguments{
\item{app}{script where shyni app is located}
}
\description{
A function to launch shiny apps
}
\examples{
\dontrun{
kobo_shiny(appname)
}


}
\author{
Elliott Messeiller
}
