% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_submission_count}
\alias{kobo_submission_count}
\title{Retrieve the number of submissions in a specified dataset}
\usage{
kobo_submission_count(formid, user = NULL, api = "kobo")
}
\arguments{
\item{formid}{The ID of the form to be accessed (as a character string).}

\item{user}{Optional. A single string indicating the username and password ("username:password"), or a character vector or list, length 2, with the first value being the "username", and the second being the "password".}

\item{api}{The URL at which the API can be accessed. Defaults to "kobo", which loads the KoBo Toolbox API.}
}
\value{
A single number indicating the number of submissions received.
}
\description{
Retrieves the number of submissions made to a specified dataset.
}
\examples{
kobo_submission_count("15051")
kobo_submission_count("31511", api = "kobohr")

}
\author{
Ananda Mahto
}

