% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kobo_time_parser_UTC}
\alias{kobo_time_parser_UTC}
\title{Parses Dates from KoBo Into a More Usable Format}
\usage{
kobo_time_parser_UTC(instring)
}
\arguments{
\item{instring}{A date/time format coming from KoBo.}
}
\value{
The \code{kobo_time_parser_UTC} function returns a POSIXct object,
while the \code{kobo_time_parser} function returns a formatted character
string that can be easily parsed as a date/time object.
}
\description{
The date/time values in KoBo usually get stored in a format
like the following: "2015-08-27T13:28:29.000+06:30". These functions
process these date/times into more usable formats.
}
\examples{
\dontrun{
TIME <- "2015-08-27T13:28:29.000+06:30"
kobo_time_parser_UTC(TIME)
}

}
\author{
Ananda Mahto
}
