% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_to_xlsform.R
\name{kobo_to_xlsform}
\alias{kobo_to_xlsform}
\title{Generate xlsfrom skeleton from a dataframe}
\usage{
kobo_to_xlsform(df, form = "form.xls", n = 100)
}
\arguments{
\item{df}{dataset to use}

\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{n}{number of levels for a factor to be considered as a text}
}
\description{
Creates and save a xlsform skeleton from a data.frames in your data folder
The form.xls will be saved in the data folder of your project.
The generated xlsfrom will need to be manually edited to configure your analysis

Note that this function only works with \code{data.frames}. The function
will throw an error for any other object types.
}
\examples{
\dontrun{
data(iris)
str(iris)
kobo_to_xlsform(iris)
}

}
\author{
Edouard Legoupil
}
