% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_trend_report.R
\name{kobo_trend_report}
\alias{kobo_trend_report}
\title{Generate trend report}
\usage{
kobo_trend_report(frame, dico)
}
\arguments{
\item{dico}{Generated from kobo_dico function}

\item{kobo}{or odk dataset to use}
}
\description{
Generate report with data aggregated by location & time
}
\examples{
kobo_trend_report(frame, dico)

\dontrun{
kobo_trend_report(frame, dico)
}

}
\author{
Edouard Legoupil
}
