% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style.R
\name{kobo_unhcr_style}
\alias{kobo_unhcr_style}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling
}
\examples{
kobo_unhcr_style()

}
\author{
Edouard Legoupil - with inspiration from bbc
}
