% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_bar.R
\name{kobo_unhcr_style_bar}
\alias{kobo_unhcr_style_bar}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style_bar()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for bar chart
}
\examples{
kobo_unhcr_style_bar()

}
\author{
Edouard Legoupil - with inspiration from bbc
}
