% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_bar_big.R
\name{kobo_unhcr_style_bar_big}
\alias{kobo_unhcr_style_bar_big}
\title{UNHCR ggplot2 theme with big labels for powerpoints}
\usage{
kobo_unhcr_style_bar_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for bar chart
}
\examples{
\dontrun{
kobo_unhcr_style_bar_big()
}

}
\author{
Edouard Legoupil - with inspiration from bbc
}
