% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_histo.R
\name{kobo_unhcr_style_histo}
\alias{kobo_unhcr_style_histo}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style_histo()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for histogram
}
\examples{
kobo_unhcr_style_histo()

}
\author{
Edouard Legoupil - with inspiration from bbc
}
