% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_histo_big.R
\name{kobo_unhcr_style_histo_big}
\alias{kobo_unhcr_style_histo_big}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style_histo_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for histogram with big labels for powerpoints
}
\examples{
kobo_unhcr_style_histo_big()

}
\author{
Edouard Legoupil - with inspiration from bbc
}
