% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_map_big.R
\name{kobo_unhcr_style_map_big}
\alias{kobo_unhcr_style_map_big}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style_map_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for maps
}
\examples{
\dontrun{
kobo_unhcr_style_map_big()
}

}
\author{
Edouard Legoupil -
}
