% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_scatter.R
\name{kobo_unhcr_style_scatter}
\alias{kobo_unhcr_style_scatter}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style_scatter()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for scatter plot
}
\examples{
kobo_unhcr_style_scatter()

}
\author{
Edouard Legoupil - with inspiration from bbc
}
