% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_unhcr_style_scatter_big.R
\name{kobo_unhcr_style_scatter_big}
\alias{kobo_unhcr_style_scatter_big}
\title{UNHCR ggplot2 theme}
\usage{
kobo_unhcr_style_scatter_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for scatter plot with big labels for powerpoints
}
\examples{
\dontrun{
kobo_unhcr_style_scatter_big()
}

}
\author{
Edouard Legoupil - with inspiration from bbc
}
