% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboloadeR-package.R
\docType{package}
\name{koboloadeR-package}
\alias{koboloadeR}
\alias{koboloadeR-package}
\title{koboloadeR: A Metapackage for Survey Data Crunching}
\description{
An organized workflow generating 'Rmd' files from an extended 'xlsform' questionnaire structure to facilitate survey data crunching.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://unhcr.github.io/koboloadeR/docs/}
  \item Report bugs at \url{https://github.com/unhcr/koboloadeR/issues}
}

}
\author{
\strong{Maintainer}: Edouard Legoupil \email{legoupil@unhcr.org}

Authors:
\itemize{
  \item Maher Daoud \email{daoudma@unhcr.org}
  \item Hisham Galal \email{galalh@unhcr.org}
}

Other contributors:
\itemize{
  \item Elliott Messeiller \email{elliott.messeiller@acted.org} [contributor]
  \item Damien Seite \email{damien.seite@eleve.ensai.fr} [contributor]
}

}
\keyword{internal}
