% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{ltbl}
\alias{ltbl}
\title{Helper function to extract the last part of question headings}
\usage{
ltbl(x, y, z)
}
\arguments{
\item{x}{= database name}

\item{y}{= column index group}

\item{z}{= column index}
}
\value{
last part of question headings
}
\description{
Helper function to extract the last part of question headings
}
\examples{
\dontrun{
ltbl(x,y,z)
}

}
\author{
Someone
}
