% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{multresponse}
\alias{multresponse}
\title{Helper function to concatenate multiple choices (select_mutiple type question) formatted TRUE / FALSE}
\usage{
multresponse(x)
}
\arguments{
\item{x}{String}
}
\value{
last part of question headings
}
\description{
Helper function to concatenate multiple choices (select_mutiple type question) formatted TRUE / FALSE
}
\examples{
\dontrun{
multresponse(x)
}

}
\author{
Someone
}
