% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{psum}
\alias{psum}
\title{Helper function that will sum values even if we have NA}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{\dots}{List of integer or numeric}
}
\value{
Integer or numeric.
}
\examples{
psum()

}
\author{
Someone
}

