% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{pwd_parse}
\alias{pwd_parse}
\title{Helper function to parse a string to be used as a username/password combination}
\usage{
pwd_parse(...)
}
\arguments{
\item{\dots}{A single string, character vetor, or list containing the username and password that should be used. If it is a single string, it should be in the form of "username:password".}
}
\description{
Converts a string of length 1 or of length 2 into a list that can then be passed on to the \code{authenticate} function from the "httr" package.
}
\note{
This function is not intended to be called directly. It is used in other functions.
}
\examples{

pwd_parse("username", "password")
pwd_parse("username:password")
pwd_parse(c("username", "password"))

}
\author{
Ananda Mahto
}

