% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Helper function to round and preserve sum}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{= value}

\item{digits}{= digits}
}
\value{
value
}
\description{
Helper function to round and preserve sum
}
\examples{
\dontrun{
round_preserve_sum(x,digits)
}

}
\author{
Someone http://biostatmatt.com/page/5
}
