% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_anonymisation_report.R
\name{kobo_anonymisation_report}
\alias{kobo_anonymisation_report}
\title{Generate a report displaying disclosure risk for Statistical Disclosure Control}
\usage{
kobo_anonymisation_report(frame, form = "form.xls", app = "console")
}
\arguments{
\item{frame}{kobo or odk dataset to use}

\item{form}{The full filename of the form to be accessed (xls or xlsx file).}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\description{
Automatically produce a disclosure risk measurement report.

 The report is generated from functions released within sdcmicro package from the worldbank.
  https://cran.r-project.org/web/packages/sdcMicro/sdcMicro.pdf
}
\examples{
\dontrun{
kobo_anonymisation_report(frame, form = "form.xls")
}

}
\author{
Edouard Legoupil
}
