% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_atlas_report.R
\name{kobo_atlas_report}
\alias{kobo_atlas_report}
\title{Generate an atlas out of the dataset}
\usage{
kobo_atlas_report(
  form = "form.xls",
  app = "console",
  output = "html",
  render = "TRUE",
  lang = "eng"
)
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}

\item{output}{The output format html or aspx if you need to upload on sharepoint), docx (to quickly cut non interesting vz and take note during data interpretation session), pptx (to quickly cut non interesting vz and persent during data interpretation session), Default is html}

\item{render}{TRUE or FALSE - Tells wheter to only produce Rmd or to also knit it in the required output format. Default is TRUE. Usefull for testing as rending takes time.}

\item{lang}{eng, fre or esp - Change the langauge of the intro to the report - default is english}
}
\value{
No return, All results will be saved on RMD files and Word files
}
\description{
Generate report with data aggregated by location & spatial visualisation / cartography
}
\examples{
\dontrun{
kobo_atlas_report(form, mapfile)
}

}
\author{
Edouard Legoupil#'
}
