% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_clean.R
\name{kobo_clean}
\alias{kobo_clean}
\title{Add cleaned variables to the frame based on a reference table}
\usage{
kobo_clean(frame, dico)
}
\arguments{
\item{dico}{Generated from kobo_dico function}

\item{kobo}{or odk dataset to use}
}
\description{
The function works on a loop based on the dictionnary.
Add a column clean and insert in the cell the name of the csv file that will be used to generate the cleaned variable
The first column of that file will be used for the matching, the second column will be added to the dataframe.
The new cleaned variable will be inserted in the dictionnary. with a suffix '.clean'
}
\examples{
kobo_clean()

\dontrun{
kobo_clean(frame, dico)
}


}
\author{
Edouard Legoupil
}
