% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_cluster_report.R
\name{kobo_cluster_report}
\alias{kobo_cluster_report}
\title{Generate anonymised dataset}
\usage{
kobo_cluster_report(frame, dico)
}
\arguments{
\item{dico}{Generated from kobo_dico function}

\item{kobo}{or odk dataset to use}
}
\description{
Automatically produce a report exploring potential clusters within survey records.

 The report is generated from functions released within FactoMiner & FactoMineR
}
\examples{
kobo_anonymise()

\dontrun{
kobo_cluster_report(frame, dico)
}

}
\author{
Edouard Legoupil
}
