% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_cluster_report.R
\name{kobo_cluster_report}
\alias{kobo_cluster_report}
\title{Generate reports with various clusterisation techniques}
\usage{
kobo_cluster_report(frame, form = "form.xls", app = "console")
}
\arguments{
\item{frame}{kobo or odk exported dataset to use}

\item{form}{The full filename of the form to be accessed (has to be an xls file).}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\description{
Automatically produce a report exploring potential clusters within survey records.

 The report is generated from functions released within FactoMiner & FactoMineR
}
\examples{
\dontrun{
kobo_cluster_report(frame =  MainDataFrame , form = "form.xls")
}

}
\author{
Edouard Legoupil
}
