% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunching_report.R
\name{kobo_crunching_report}
\alias{kobo_crunching_report}
\title{Generate Data Crunching Report}
\usage{
kobo_crunching_report(form = "form.xls", app = "console")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}

\item{app}{The place where the function has been executed, the default is the console and the second option is the shiny app}
}
\value{
No return, All results will be saved on RMD files and Word files
}
\description{
Generate crunching Report that contains all descriptive statistics, correlation analysis, tabulation and data visualization for variables and indicators.
}
\examples{
kobo_crunching_report()

\dontrun{
kobo_crunching_report("myform.xls")
}

}
\author{
Edouard Legoupil, Maher Daoud
}
