% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_split_multiple.R
\name{kobo_split_multiple}
\alias{kobo_split_multiple}
\title{Split variables resulting from select_multiple questions}
\usage{
kobo_split_multiple(data, dico)
}
\arguments{
\item{data}{Dataframe with selectmultiple column to split}

\item{dico}{Data dictionnary generated from kobo_dico}
}
\value{
data A "data.table" with the full splitted select_multiple.
}
\description{
To be used when extracting from ODK that does not offers splitting capacity
}
\examples{
\dontrun{
kobo_split_multiple(data, dico)
}

}
\author{
Edouard Legoupil
}
