% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_to_xlsform.R
\name{kobo_to_xlsform}
\alias{kobo_to_xlsform}
\title{Generate xlsfrom skeleton from a dataframe}
\usage{
kobo_to_xlsform(df, n = 100)
}
\arguments{
\item{n}{number of levels for a factor to be considered as a text}
}
\description{
Creates and save a xlsform skeleton from a data.frames in your data folder
The form.xls will be saved in the data folder of your project.
The generated xlsfrom will need to be manually edited to configure your analysis

Note that this function only works with \code{data.frames}. The function
will throw an error for any other object types.
}
\examples{
data(iris)
str(iris)
kobo_to_xlsform(iris)
}
\author{
Edouard Legoupil
}
