% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{psum}
\alias{psum}
\title{Sum with NA}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{\dots}{List of integer or numeric}

\item{na.rm}{Bolean indicating if NA shall be removed}
}
\value{
Integer or numeric.
}
\description{
Helper function that will sum values even if we have NA
}
\examples{
\dontrun{
psum()
}

}
\author{
Someone
}
