% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{round2}
\alias{round2}
\title{Create roundup function}
\usage{
round2(x, n)
}
\arguments{
\item{List}{of integer or numeric to be rounded}

\item{Rounding}{level}
}
\value{
rounded figure.
}
\description{
Create roundup function
}
\examples{
round2(x, n)

}
\author{
Someone
}
